requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
markdownify==0.13.1
numpy==1.26.3
outcome==1.3.0.post0
pandas==2.2.3
peewee==3.17.9
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
#python-jobspy==1.1.76
python-telegram-bot==21.10
pytz==2025.1
regex==2024.11.6
requests==2.32.3
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tls-client==1.0.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
